Rename manager to base
