+++ e2xauthoring/managers/presetmanager.py
@@ -3,7 +3,7 @@ import os
 import nbformat
 from traitlets import Unicode
 
-from .manager import BaseManager
+from .base import BaseManager
 
 
 class PresetManager(BaseManager):
