+++ e2xauthoring/managers/worksheetmanager.py
@@ -4,7 +4,7 @@ import shutil
 
 from traitlets import Unicode
 
-from .manager import BaseManager
+from .base import BaseManager
 
 
 class WorksheetManager(BaseManager):
